emotion_detection_model/requirements_server.txt
flask
flask-socketio
torch
torchvision
opencv-python
numpy
python-socketio
Pillow
eventlet
gunicorn